changed cluster.pl
This version of cluster.pl first groups reads by SbfI position and
then clusters within groups based on the sequence upstream of SbfI.
This is to dertermine whether repeat sequences are responsible for
for the spike in frequency of reads having and SbfI site at the end.
If the Sbfi site is not at the end of the read, then genomic
religation diversifies the sequence downstream of the SbfI site, which
generally prevents clustering.